Inline registry tests
